+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 
 - Added status indicator when sending a post (#7)
 - Added alert when sending/editing post/comment fails
+- Refreh button in the inbox
 
 ### Bug Fixes
 
 - Fixed notification counter not updating (#9)
+- Unread message count should update, as soon as you reply to a message from the inbox
 
 ## 0.1 (2023-07-17)